Module creating and their dependencies are refactored completely, everything OK now...
