+++ work_manager/src/main/java/com/work_manager/NewsWorkManager.kt
@@ -2,7 +2,6 @@ package com.work_manager
 
 import android.content.Context
 import androidx.work.*
-import data.workManager.UploadWorker
 import java.util.*
 import java.util.concurrent.TimeUnit
